backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
PyPDF2==3.0.1
pdfplumber==0.10.3
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
openai==1.54.3
python-dotenv>=1.0.0
numpy>=1.24.3,<2.0.0
huggingface-hub>=0.19.4
transformers>=4.35.2
torch>=2.1.0
